Adjust hover colours and use generic DataTable component
