+++ editor.planx.uk/src/pages/PlatformAdminPanel/PlatformAdminPanel.tsx
@@ -2,1348 +2,23 @@ import Box from "@mui/material/Box";
 import Container from "@mui/material/Container";
 import Typography from "@mui/material/Typography";
 import {
-  DataGrid,
   GridColDef,
   GridRenderCellParams,
   GridTreeNodeWithRender,
 } from "@mui/x-data-grid";
+import { useStore } from "pages/FlowEditor/lib/store";
 import React from "react";
 import { AdminPanelData } from "types";
 import SettingsSection from "ui/editor/SettingsSection";
+import { DataTable } from "ui/shared/DataTable/DataTable";
 
 import { Article4Status } from "./components/Article4Status";
 import { Configured, NotConfigured } from "./components/icons";
 
 export const PlatformAdminPanel = () => {
-  // const adminPanelData: AdminPanelData[] | undefined = useStore(
-  //   (state) => state.adminPanelData,
-  // );
-
-  // console.log({adminPanelData})
-
-  const adminPanelData: AdminPanelData[] = [
-    {
-      __typename: "teams_summary",
-      id: 20,
-      name: "Barking & Dagenham",
-      slug: "barking-and-dagenham",
-      referenceCode: "BDG",
-      homepage: "https://www.lbbd.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Article4 barkinganddagenham",
-        "Article4 barkinganddagenham (no Becontree)",
-        "Backup dec23 article4 barkinganddagenham manual constraints",
-        "Backup dec23 becontree design code",
-        "Backup dec23 find out if you need planning permission",
-        "Backup feb24 article4 barkinganddagenham manual constraints",
-        "Backup jan24 find out if you need planning permission testing manual becontree",
-        "Backup mar19 find out if you need planning permission testing pre infill",
-        "Becontree design code",
-        "Find out if you need planning permission (no Becontree)",
-        "Find out if you need planning permission testing",
-        "Find out what planning contraints apply to a property",
-        "Retrofit guide",
-        "Retrofit nudges",
-        "Select the works full",
-        "Test",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/barking-and-dagenham.svg",
-      favicon: null,
-      primaryColour: "#0B0C0C",
-      linkColour: "#0B0C0C",
-      actionColour: "#0B0C0C",
-    },
-    {
-      __typename: "teams_summary",
-      id: 26,
-      name: "Barnet",
-      slug: "barnet",
-      referenceCode: "BNE",
-      homepage: "https://barnet.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "planxenquiries@barnet.gov.uk",
-        helpPhone: "020 8359 3000",
-        emailReplyToId: "f7292839-4f07-4938-aa90-763ebf968a1a",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: true,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://barnet.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Apply for a lawful development certificate",
-        "Apply for planning advice barnet template",
-        "Apply for planning permission",
-        "Apply for prior approval",
-        "Apply for service declarations",
-        "Article4",
-        "Example submission service",
-        "Find out if you need planning permission",
-        "Householder Barnet validation requirements",
-        "Initial planning advice barnet",
-        "Pre application planning advice",
-        "Report a planning breach",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/barnet.svg",
-      favicon: null,
-      primaryColour: "#007078",
-      linkColour: "#007078",
-      actionColour: "#007078",
-    },
-    {
-      __typename: "teams_summary",
-      id: 22,
-      name: "Birmingham",
-      slug: "birmingham",
-      referenceCode: "BIR",
-      homepage: "https://birmingham.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: ["Article4", "Find out if you need planning permission"],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/birmingham.png",
-      favicon: null,
-      primaryColour: "#84329B",
-      linkColour: "#84329B",
-      actionColour: "#84329B",
-    },
-    {
-      __typename: "teams_summary",
-      id: 9,
-      name: "Braintree",
-      slug: "braintree",
-      referenceCode: "BRA",
-      homepage: "https://www.braintree.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: false,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "planning@braintree.gov.uk",
-        helpPhone: "01376 557779",
-        emailReplyToId: "9fa4a03d-f981-4eb7-b7f7-1a1842321efc",
-        helpOpeningHours: "Mon-Fri 9am-5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: ["Article4", "Find out if you need planning permission"],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/braintree.svg",
-      favicon: null,
-      primaryColour: "#332e35",
-      linkColour: "#332e35",
-      actionColour: "#332e35",
-    },
-    {
-      __typename: "teams_summary",
-      id: 7,
-      name: "Bromley",
-      slug: "bromley",
-      referenceCode: "BRY",
-      homepage: "https://www.bromley.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: false,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: ["Article4"],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/bromley.svg",
-      favicon: null,
-      primaryColour: "#019245",
-      linkColour: "#019245",
-      actionColour: "#019245",
-    },
-    {
-      __typename: "teams_summary",
-      id: 4,
-      name: "Buckinghamshire",
-      slug: "buckinghamshire",
-      referenceCode: "BKM",
-      homepage: "https://www.buckinghamshire.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "planning.digital@buckinghamshire.gov.uk",
-        helpPhone: "0300 131 6000",
-        emailReplyToId: "8a28d708-3599-4fcc-82bf-c73d15e54acd",
-        helpOpeningHours: "Monday to Thursday 9am to 5:30pm, Friday 9am to 5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://buckinghamshire.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Amend property outline buckinghamshire",
-        "Apply for a lawful development certificate",
-        "Apply for a lawful development certificate template (copy)",
-        "Apply for building control conservatories and porches exempt bc",
-        "Apply for building regulations applications",
-        "Apply for building regulations full plans",
-        "Apply for building regulations notice",
-        "Apply for building regulations notice replacement windows or doors",
-        "Apply for ldc planningportal simulation",
-        "Apply for listed building advice",
-        "Apply for planning permission",
-        "Apply for planning permission (Idox testing)",
-        "Apply for prior approval",
-        "Apply for prior approval (Idox testing)",
-        "Apply for regularisation",
-        "Article4",
-        "Check whether you need a building control application",
-        "Check whether your project is exempt from building regulations",
-        "Digital planning uniform change request",
-        "Find out if you need planning permission",
-        "Get copies of building regulations documents",
-        "Initial planning advice",
-        "Notification of commencement",
-        "Notify us of any high efficiency alternative energy systems",
-        "Planning advice",
-        "Planning history check",
-        "Planning tech and business support planx",
-        "Planning tech and business support planx v2",
-        "Report a high hedge",
-        "Report a planning breach",
-        "Report a planning breach v1superseded",
-        "Report a planning breach working group v2",
-        "Report a potential dangerous structure",
-        "Request a building control quote",
-        "Request a building control site inspection",
-        "Request an ecology screening visit",
-        "Resume a previous building control application",
-        "Submit a demolition notice",
-        "Temporary pay test",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/buckinghamshire.svg",
-      favicon: null,
-      primaryColour: "#212D77",
-      linkColour: "#212D77",
-      actionColour: "#212D77",
-    },
-    {
-      __typename: "teams_summary",
-      id: 18,
-      name: "Camden",
-      slug: "camden",
-      referenceCode: "CMD",
-      homepage: "https://camden.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "planning@camden.gov.uk",
-        helpPhone: "020 7974 4444",
-        emailReplyToId: "009d11b3-7448-4699-a6d3-a5aff8acab59",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://camden.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Apply for listed building consent",
-        "Apply for prior approval",
-        "Article4",
-        "Article4 copy",
-        "CAMDEN - Apply for planning permission",
-        "Dsn impact metrics",
-        "Find out if you need planning permission",
-        "Find out if you need planning permission - Energy efficiency",
-        "Housing flags",
-        "Pay fee calculation example",
-        "Pre app",
-        "Pre-application for energy efficient measures",
-        "Report a planning breach camden",
-        "Spatial data component bundle",
-        "Test apply for prior approval",
-        "TESTING - Amend a planning permission template",
-        "TESTING - Apply for listed building consent template",
-        "TESTING - Apply for pre-application advice",
-        "TESTING - Report a planning breach working group v2.1",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/camden.svg",
-      favicon:
-        "https://api.editor.planx.uk/file/public/hhvhhip7/Untitled%20design%20-%202024-03-04T133903.500.png",
-      primaryColour: "#000000",
-      linkColour: "#000000",
-      actionColour: "#000000",
-    },
-    {
-      __typename: "teams_summary",
-      id: 2,
-      name: "Canterbury",
-      slug: "canterbury",
-      referenceCode: "CAT",
-      homepage: "https://www.canterbury.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: ["Article4check", "Search"],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/canterbury.svg",
-      favicon: null,
-      primaryColour: "#331035",
-      linkColour: "#331035",
-      actionColour: "#331035",
-    },
-    {
-      __typename: "teams_summary",
-      id: 14,
-      name: "Doncaster",
-      slug: "doncaster",
-      referenceCode: "DNC",
-      homepage: "https://doncaster.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "tsi@doncaster.gov.uk",
-        helpPhone: "(01302) 734854",
-        emailReplyToId: "0561276a-a2bd-4c31-b83b-ae716f53f6eb",
-        helpOpeningHours: "Monday - Friday, 8.30am - 5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: true,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Amend a planning permission template (copy)",
-        "Apply for a lawful development certificate ",
-        "Apply for a rights of way order (copy)",
-        "Apply for planning permission (testing copy)",
-        "Apply for pre application advice (V1 Doncaster service)",
-        "Article4",
-        "Find out if you need planning permission",
-        "Listed building pre application advice",
-        "Notify commencement of planning application",
-        "Notify completion of planning application",
-        "Pay fee calculation example",
-        "Pre application advice",
-        "Report a planning breach",
-        "Tree pre application advice",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/doncaster.svg",
-      favicon: null,
-      primaryColour: "#981d6a",
-      linkColour: "#981d6a",
-      actionColour: "#981d6a",
-    },
-    {
-      __typename: "teams_summary",
-      id: 33,
-      name: "Environment Agency",
-      slug: "environment-agency",
-      referenceCode: null,
-      homepage: null,
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: null,
-      logo: "https://api.editor.planx.uk/file/public/twomnxfe/EA_logo.svg",
-      favicon: "https://api.editor.planx.uk/file/public/durm8ner/Favicon.png",
-      primaryColour: "#00803d",
-      linkColour: "#00803d",
-      actionColour: "#00803d",
-    },
-    {
-      __typename: "teams_summary",
-      id: 21,
-      name: "Epsom & Ewell",
-      slug: "epsom-and-ewell",
-      referenceCode: "EPS",
-      homepage: "https://www.epsom-ewell.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "planningsupport@epsom-ewell.gov.uk",
-        helpPhone: "(01372) 732000",
-        emailReplyToId: "493d0dc1-4aa7-400a-b092-2953ca8b04c0",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: true,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://epsom-ewell.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Apply for a Lawful Development Certificate ",
-        "Article4",
-        "Article4old",
-        "Find out if you need planning permission",
-        "Report a planning breach ",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/epsom-and-ewell.svg",
-      favicon: null,
-      primaryColour: "#2F4538",
-      linkColour: "#2F4538",
-      actionColour: "#2F4538",
-    },
-    {
-      __typename: "teams_summary",
-      id: 27,
-      name: "Gateshead",
-      slug: "gateshead",
-      referenceCode: "GAT",
-      homepage: "https://gateshead.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "enforcement@gateshead.gov.uk",
-        helpPhone: "0191 433 7225",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Article4",
-        "Find out if you need planning permission",
-        "Let the planning team know about an issue",
-        "Report a planning breach ",
-        "Report a potential dangerous structure template copy",
-        "Test",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/gateshead.svg",
-      favicon: null,
-      primaryColour: "#0A0A0A",
-      linkColour: "#0A0A0A",
-      actionColour: "#0A0A0A",
-    },
-    {
-      __typename: "teams_summary",
-      id: 19,
-      name: "Gloucester",
-      slug: "gloucester",
-      referenceCode: "GLO",
-      homepage: null,
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: true,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://gloucester.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Amend a planning permission template (copy)",
-        "Apply for listed building consent",
-        "Apply for planning permission",
-        "Apply for pre application advice (copy)",
-        "Apply for pre application advice (V1 Gloucester service)",
-        "Apply for prior approval",
-        "Article4",
-        "Check for restrictive conditions",
-        "Check if you need planning permission",
-        "Report a planning breach working group v2",
-      ],
-      logo: "https://api.editor.planx.uk/file/public/n9snbmgv/gcc-white-tag-logo.svg",
-      favicon: "https://api.editor.planx.uk/file/public/5s2llxpn/favicon.png",
-      primaryColour: "#0054a4",
-      linkColour: "#0054a4",
-      actionColour: "#0054a4",
-    },
-    {
-      __typename: "teams_summary",
-      id: 31,
-      name: "Kingston",
-      slug: "kingston",
-      referenceCode: "KTT",
-      homepage: "https://www.kingston.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: false,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: ["Article4", "Find out if you need planning permission"],
-      logo: null,
-      favicon: null,
-      primaryColour: "#0010A4",
-      linkColour: "#0010A4",
-      actionColour: "#0010A4",
-    },
-    {
-      __typename: "teams_summary",
-      id: 3,
-      name: "Lambeth",
-      slug: "lambeth",
-      referenceCode: "LBH",
-      homepage: "https://lambeth.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "digitalplanning@lambeth.gov.uk",
-        helpPhone: "020 7926 9955",
-        emailReplyToId: "b1638bb8-a16b-4c21-955a-5ec84f65b976",
-        helpOpeningHours: "Mon-Fri 9am-5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: true,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://lambeth.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Accessibility test",
-        "Apply for a lawful development certificate",
-        "Apply for a lawful development certificate (Idox testing)",
-        "Apply for planning permission",
-        "Apply for prior approval",
-        "Check constraints on a property",
-        "Document guidance user testing",
-        "Fin planning wizard amazing test service",
-        "Find out if you need planning permission",
-        "Lambeth article 4 check",
-        "Options for dealing with planning judgment",
-        "Submit details required by a condition",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/lambeth.svg",
-      favicon:
-        "https://api.editor.planx.uk/file/public/t2xea2fx/Lambethfavicon.png",
-      primaryColour: "#006360",
-      linkColour: "#006360",
-      actionColour: "#006360",
-    },
-    {
-      __typename: "teams_summary",
-      id: 16,
-      name: "Medway",
-      slug: "medway",
-      referenceCode: "MDW",
-      homepage: "https://medway.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "customerplanning@medway.gov.uk",
-        helpPhone: "01634 331700",
-        emailReplyToId: "a9f18954-800c-4236-ba2f-c2f6eb4769fe",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://medway.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Apply for a lawful development certificate",
-        "Apply for listed building consent",
-        "Apply for planning permission",
-        "Apply for prior approval",
-        "Article4",
-        "Dummy test flow",
-        "Find out if you need planning permission",
-        "Householder Medway validation requirements ",
-        "Notice of exempt works to trees - Final Checks",
-        "Report a derelict or empty building",
-        "Report a planning breach",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/medway.svg",
-      favicon: null,
-      primaryColour: "#172b4e",
-      linkColour: "#172b4e",
-      actionColour: "#172b4e",
-    },
-    {
-      __typename: "teams_summary",
-      id: 15,
-      name: "Newcastle",
-      slug: "newcastle",
-      referenceCode: "NET",
-      homepage: "https://newcastle.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://newcastle.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "Apply for planning permission",
-        "Article4",
-        "Ask a question knowledge base",
-        "Find out if you need planning permission",
-        "Householder files copy for newcastle",
-        "Nb temp",
-        "Nb test constraints",
-        "Nb validation requirements",
-        "Nb validation requirements copy",
-        "Newcastle works to trees filter",
-        "Validation requirements check",
-        "Validation requirements check copy",
-        "Validation requirements check live",
-        "Validation requirements check working copy hh comp",
-        "Validation requirements check workingversion",
-        "Validation requirements (live)",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/newcastle.svg",
-      favicon: null,
-      primaryColour: "#000000",
-      linkColour: "#000000",
-      actionColour: "#000000",
-    },
-    {
-      __typename: "teams_summary",
-      id: 30,
-      name: "Open Digital Planning",
-      slug: "open-digital-planning",
-      referenceCode: null,
-      homepage: "https://opendigitalplanning.org/",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "93e82367-c66c-4043-831c-be22af1987bd",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Apply for listed building consent",
-        "Apply for planning permission July 2024 User Research (DEPRECATED & CORRUPTED)",
-        "Calculate biodiversity net gain",
-        "Find out if you need planning permission",
-        "Report a planning breach about the breach",
-        "Report a planning breach scoping",
-        "Report a planning breach scoping bucks SUPERSEDED",
-        "Validating and assessing applications with heritage impacts",
-      ],
-      logo: null,
-      favicon: null,
-      primaryColour: "#0010A4",
-      linkColour: "#0010A4",
-      actionColour: "#0010A4",
-    },
-    {
-      __typename: "teams_summary",
-      id: 1,
-      name: "Open Systems Lab",
-      slug: "opensystemslab",
-      referenceCode: null,
-      homepage: null,
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "About the applicant",
-        "Amend property outline",
-        "Apply for planning permission ownership certificates",
-        "Apply for planning permission service template",
-        "Apply for planning permission triage all types",
-        "Apply for planning permission triage (old)",
-        "Apply for pp householder guidance",
-        "Apply for service declarations",
-        "Article4 template",
-        "Automation test",
-        "Biodiversitynetgain triage",
-        "Calculate fee test service",
-        "Caravans and mobile homes",
-        "Certificate of lawfulness documents immunity",
-        "Certificate of lawfulness existing drawings",
-        "Certificate of lawfulness existing drawings immunity",
-        "Certificate of lawfulness existing drawings immunity old",
-        "Certificate of lawfulness existing drawings old",
-        "Certificate of lawfulness proposed drawings",
-        "Certificate of lawfulness proposed drawings old",
-        "Change the use of a property class a",
-        "Change the use of a property class aa",
-        "Change the use of a property class d",
-        "Change the use of a property class g",
-        "Change the use of a property class h",
-        "Change the use of a property class l",
-        "Change the use of a property class m",
-        "Change the use of a property class ma",
-        "Change the use of a property class n",
-        "Change the use of a property class q",
-        "Change the use of a property class r ",
-        "Change the use of a property class s",
-        "Change the use of a property class t",
-        "Change the use of a property home to hmo",
-        "Change the use of a property in the future (deprecated)",
-        "Change the use of a property in the past",
-        "Change the use of a property test",
-        "Change the use of a property work from home",
-        "Change the use of land in the future",
-        "Change the use of land in the past",
-        "Community infrastructure levy",
-        "Content design guidance",
-        "Design and access statement form",
-        "Digital land api search test",
-        "Fairhold",
-        "Fee calculator",
-        "Fee calculator info only",
-        "Find out if you need planning permission template",
-        "Fire statement template",
-        "Flood risk assesment template keep and improve",
-        "Foiynpp skip q draft",
-        "Foiynpp skip q draft copy",
-        "Heritage statement form",
-        "Householder files",
-        "Householder files old",
-        "Householder more information",
-        "Immunity check",
-        "In progress permitted development",
-        "Lawful development certificate more details",
-        "Ldc e non compliance with a planning condition",
-        "Listed building consent confirmation pages",
-        "Listed building consent files",
-        "Listed building consent more information",
-        "Listed building consent simple v1 triage",
-        "London data hub (old)",
-        "Materials (old)",
-        "Minor major files",
-        "Minor major more information (old)",
-        "Modules permitted develolpment",
-        "National test apply for planning permission",
-        "National test find out if you need planning permission",
-        "Material and non material amendments scoping",
-        "Ownership",
-        "Parent shell filter",
-        "Pd restrictions information",
-        "Pd restructure",
-        "Permitted development work from home",
-        "Permitteddevelopment",
-        "Permitteddevelopment add homes on detached commercial buildings",
-        "Permitteddevelopment add homes onto detached flats",
-        "Permitteddevelopment add homes onto detached homes",
-        "Permitteddevelopment add homes onto terraced commercial mixed used building",
-        "Permitteddevelopment add homes onto terraced homes",
-        "Permitteddevelopment add storeys to a house",
-        "Permitteddevelopment agricultural buildings",
-        "Permitteddevelopment demolish",
-        "Permitteddevelopment demolish vacant buildings and replace with homes",
-        "Permitteddevelopment new forestry building",
-        "Permitteddevelopment rear and side extensions",
-        "Permitteddevelopment solar panels",
-        "Pre app advice check",
-        "Prior approval assessment information",
-        "Prior approval confirmation pages",
-        "Prior approval files",
-        "Prior approval files old",
-        "Prior approval more information",
-        "Prior approval scope check",
-        "Property types",
-        "Regularise use",
-        "Residential units (old)",
-        "Select the works",
-        "Select the works triage",
-        "Small sites metric tool",
-        "Spatial data component bundles",
-        "Uniform translator",
-        "Works to trees files",
-        "Works to trees filter",
-      ],
-      logo: null,
-      favicon: null,
-      primaryColour: "#0010A4",
-      linkColour: "#0010A4",
-      actionColour: "#0010A4",
-    },
-    {
-      __typename: "teams_summary",
-      id: 24,
-      name: "PlanX",
-      slug: "planx",
-      referenceCode: null,
-      homepage: "https://www.planx.uk/",
-      subdomain: null,
-      planningDataEnabled: false,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "support@planx.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: "jo@opensystemslab.io",
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Confirm that you have published your spatial data",
-        "Give us contact information for govuk notify",
-        "Its georges birthday",
-        "jo",
-        "Raise a purchase order number",
-        "Send us information about your article 4 directions",
-        "Send us questions from your it team",
-        "Send us the email addresses you would like to access your planx workspace",
-        "Send us your agreement",
-        "Send us your branding information",
-        "Tell us about your team",
-        "Tell us that your it team have approved the use of plan",
-        "Tell us what services are live",
-        "Tell us where you want your applications submitted to",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/PlanX_logo_white.svg",
-      favicon: null,
-      primaryColour: "#0010A4",
-      linkColour: "#0010A4",
-      actionColour: "#0010A4",
-    },
-    {
-      __typename: "teams_summary",
-      id: 8,
-      name: "Scotland",
-      slug: "scotland",
-      referenceCode: null,
-      homepage: null,
-      subdomain: null,
-      planningDataEnabled: false,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Check if you need a building warrant",
-        "Check what permissions your building project is likely to need",
-        "Gis search test",
-        "Permitteddevelopment rear and side extensions scotland",
-        "Permitteddevelopment scotland",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/scotland.svg",
-      favicon: null,
-      primaryColour: "#0065bd ",
-      linkColour: "#0065bd ",
-      actionColour: "#0065bd ",
-    },
-    {
-      __typename: "teams_summary",
-      id: 6,
-      name: "Southwark",
-      slug: "southwark",
-      referenceCode: "SWK",
-      homepage: "https://www.southwark.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "digitalprojects@southwark.gov.uk",
-        helpPhone: "020 7525 5403",
-        emailReplyToId: "43aee528-833a-49ae-b373-174126f9e815",
-        helpOpeningHours:
-          "Monday to Friday, 9am–5.30pm (except public holidays)",
-      },
-      govpayEnabled: true,
-      powerAutomateEnabled: true,
-      sendToEmailAddress: null,
-      bopsSubmissionURL:
-        "https://southwark.bops-staging.services/api/v2/planning_applications",
-      liveFlows: [
-        "AA Test run",
-        "Amend a planning permission template",
-        "Apply for a lawful development certificate",
-        "Apply for building regulations full plans",
-        "Apply for listed building consent ",
-        "Apply for planning permission",
-        "Apply for prior approval",
-        "Apply for works to trees",
-        "Article4",
-        "Call for sites",
-        "Find out if you need planning permission",
-        "Pre apps",
-        "Report a planning breach",
-        "Report a planning breach V1",
-        "Report a potential dangerous structure template",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/southwark.svg",
-      favicon: null,
-      primaryColour: "#256e8a",
-      linkColour: "#256e8a",
-      actionColour: "#256e8a",
-    },
-    {
-      __typename: "teams_summary",
-      id: 28,
-      name: "St Albans",
-      slug: "st-albans",
-      referenceCode: "SAL",
-      homepage: "https://www.stalbans.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: ["Article4", "Find out if you need planning permission"],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/st-albans.svg",
-      favicon: null,
-      primaryColour: "#039",
-      linkColour: "#039",
-      actionColour: "#039",
-    },
-    {
-      __typename: "teams_summary",
-      id: 29,
-      name: "Templates",
-      slug: "templates",
-      referenceCode: "BBD",
-      homepage: null,
-      subdomain: null,
-      planningDataEnabled: false,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Amend property outline template",
-        "Apply for a lawful development certificate template",
-        "Apply for planning permission template (copy)",
-        "Apply for service name template",
-        "Article4 template",
-        "Check constraints on a property template",
-        "Find out if you need planning permission template",
-        "Find out what planning contraints apply to a property",
-        "Generic submit information template",
-        "Notification of Completion",
-        "Submission service template",
-      ],
-      logo: null,
-      favicon: null,
-      primaryColour: "#0010A4",
-      linkColour: "#0010A4",
-      actionColour: "#0010A4",
-    },
-    {
-      __typename: "teams_summary",
-      id: 5,
-      name: "Testing",
-      slug: "testing",
-      referenceCode: null,
-      homepage: null,
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.gov.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Ab testing",
-        "Abc",
-        "Add project test",
-        "Adding a project to the projects list test",
-        "Additiona material example",
-        "Alastairs skip experiment",
-        "Alastairs skip experiment copy",
-        "Analytics tracking test",
-        "Another filter",
-        "Apply for Planning Permission",
-        "Apply for planning permission copy",
-        "Apply for planning permission old backup",
-        "Apply for planning permission service template old",
-        "Apply for planning permission triage for testing",
-        "Apply for pp householder guidance",
-        "Apply for pre application advice (V1 Camden service)",
-        "August portal one level",
-        "August portal test",
-        "Auto answer next",
-        "Auto answering the future bug example",
-        "Automate blanks test",
-        "Automations external portal",
-        "Automations test flow",
-        "Back test",
-        "Barkingdagenham a4 test",
-        "Basic",
-        "Boundary",
-        "Branch filter test",
-        "Bucks a4 test",
-        "Calcpay",
-        "Calculating fee for material change of use test",
-        "Change of use ma fix test",
-        "Change of use map",
-        "Change the use of a property in the future copy",
-        "Change tracking",
-        "Changepasttest",
-        "Checklist behaviour test",
-        "Checklist compound test",
-        "Checklist logic problem",
-        "Child test",
-        "Child test copy",
-        "Clone test service",
-        "Common allergens",
-        "Condition questions alternative",
-        "Constraints",
-        "Corrupted edge node test",
-        "Create streets",
-        "Daf test",
-        "Daf test 2",
-        "data-migration-test",
-        "Dates",
-        "Demo",
-        "Disabled button test",
-        "Do my shopping",
-        "Doors test",
-        "Double find property",
-        "Drawings test",
-        "E2e test",
-        "Empty",
-        "Error messages",
-        "External portal blank screen replication test",
-        "External portal test",
-        "Falgs",
-        "Fee calculator copy",
-        "Fee calculator presentation",
-        "Filter bugs",
-        "Filter flow test triage",
-        "Filter test nobug 2",
-        "Filter test with bug",
-        "Filters",
-        "Find out if you need planning permission",
-        "Find out if you need planning permission check pd restrictions",
-        "Find out if you need planning permission copy",
-        "Find out if you need planning permission - JC update testing",
-        "Find out if you need planning permission old template",
-        "Findprop",
-        "Flag order test",
-        "Flag order test with autoanswer",
-        "Flags example",
-        "Gis test",
-        "Hasura api tests",
-        "Hmmmm",
-        "Householder more information user testing",
-        "Human readable projects test",
-        "Immunity test 1",
-        "Inner setvalue",
-        "intermediate-value-bug",
-        "Internal works only test",
-        "Invite to pay draft",
-        "Its ians birthday",
-        "Jedi automations",
-        "Just a notice",
-        "Just some automation",
-        "Just some other automation",
-        "Ldc new warnings",
-        "Ldc without external portals",
-        "Lighthouse canary flow",
-        "Link tab test",
-        "Links test",
-        "List component profiling",
-        "List component testing",
-        "London data hub testing",
-        "Material change test",
-        "Mikes analytics testing",
-        "Minor major more information testing",
-        "Multiple journey context testing",
-        "Multiple results prototyping",
-        "National test apply for planning permission",
-        "National test apply for planning permission old",
-        "New pd check structure",
-        "New project details apply for planning permission",
-        "New uses for calculator",
-        "Not null test",
-        "Nots test",
-        "Number and calculate test",
-        "Number input zero",
-        "Officer checklist",
-        "Only test",
-        "Order a meal",
-        "Order test",
-        "Override",
-        "Overwrite test",
-        "Parent child automations test",
-        "Parent test",
-        "Pay test",
-        "Pay testing",
-        "Pay your building safety levy",
-        "Permitted development archive",
-        "Planx logic potpourri",
-        "Playground",
-        "Portal test",
-        "Prior approval files old backup",
-        "Process",
-        "Projecttype full list",
-        "Property override",
-        "Property type unit type sorting",
-        "Publishing clones",
-        "Question rearrangement and deletion",
-        "Reconciliation answers test",
-        "Recoveryexample",
-        "Report a planning breach scoping copy",
-        "Result accessibility test",
-        "Results triage bug",
-        "Results triage bug portal",
-        "Rich text test",
-        "Russian doll",
-        "S3 upload test",
-        "Sections nav test",
-        "Sections test",
-        "Select the works hh fee test",
-        "Select the works ldce fee test",
-        "Select the works ldcp fee test",
-        "Select the works major fee test",
-        "Select the works minor fee test",
-        "Select the works triage new dec 2023",
-        "Select the works triage new dec 2023 var2",
-        "Select the works triage new jan 2023 minimalist pp",
-        "Set value append replace",
-        "Set value append replace copy",
-        "Set value appending behaviour",
-        "Set value test",
-        "Sharedb demo",
-        "Ssm calc test",
-        "Stripped foi logic",
-        "Submission test",
-        "Test",
-        "Test flow copy",
-        "Test for e2e",
-        "Test for video",
-        "Test service",
-        "Test upload and label",
-        "Testing jonny",
-        "Testsst",
-        "Title boundaries",
-        "Trees only test",
-        "True false constraints",
-        "Unpublished test",
-        "Upload and label bug fix",
-        "Upload and label demo",
-        "Upload label logic example",
-        "Upload test",
-        "Uploads video",
-        "Variable nightmare",
-        "Window works drawings",
-        "Windows high contrast",
-        "Wrong project selection solution",
-      ],
-      logo: null,
-      favicon: null,
-      primaryColour: "#0010A4",
-      linkColour: "#0010A4",
-      actionColour: "#0010A4",
-    },
-    {
-      __typename: "teams_summary",
-      id: 25,
-      name: "Tewkesbury",
-      slug: "tewkesbury",
-      referenceCode: "TEW",
-      homepage: "https://tewkesbury.gov.uk",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "development.applications@tewkesbury.gov.uk",
-        helpPhone: "01684 295010",
-        emailReplyToId: "cc56b82a-3350-4331-97cb-dfa92909bf74",
-        helpOpeningHours:
-          "Monday - Thursday, 8:30am - 5pm, Friday 8:30am - 4pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Apply for a lawful development certificate play",
-        "Article4",
-        "Check if you need planning permission",
-        "Check my PD Rights",
-        "Check Your Planning Constraints ",
-        "Discharge of conditions",
-        "Drop kerbs",
-        "Historic files request",
-        "Pre application advice",
-        "Report a planning breach",
-        "Report a planning breach working group v2 copy",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/tewkesbury.svg",
-      favicon: null,
-      primaryColour: "#1d3d4f",
-      linkColour: "#1d3d4f",
-      actionColour: "#1d3d4f",
-    },
-    {
-      __typename: "teams_summary",
-      id: 23,
-      name: "West Berkshire",
-      slug: "west-berkshire",
-      referenceCode: "WBK",
-      homepage: "https://www.westberks.gov.uk/",
-      subdomain: null,
-      planningDataEnabled: true,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "planningenforcement@westberks.gov.uk",
-        helpPhone: "(01635) 551111",
-        emailReplyToId: "36d09695-3a17-4f42-9280-d6b75189642d",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Article4",
-        "Find out if you need planning permission",
-        "Report a planning breach",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/west-berkshire.svg",
-      favicon: null,
-      primaryColour: "#007046",
-      linkColour: "#007046",
-      actionColour: "#007046",
-    },
-    {
-      __typename: "teams_summary",
-      id: 17,
-      name: "WikiHouse",
-      slug: "wikihouse",
-      referenceCode: null,
-      homepage: null,
-      subdomain: null,
-      planningDataEnabled: false,
-      article4sEnabled: "@todo",
-      govnotifyPersonalisation: {
-        helpEmail: "example@council.co.uk",
-        helpPhone: "(01234) 567890",
-        emailReplyToId: "727d48fa-cb8a-42f9-b8b2-55032f3bb451",
-        helpOpeningHours: "Monday - Friday, 9am - 5pm",
-      },
-      govpayEnabled: false,
-      powerAutomateEnabled: false,
-      sendToEmailAddress: null,
-      bopsSubmissionURL: null,
-      liveFlows: [
-        "Eligibility for fast track engineering service",
-        "Eligibility for fast track engineering service 3",
-        "Eligibility for services",
-        "Start your project",
-        "Start your project 2",
-        "Start your project 3",
-        "Start your project 4 by ar",
-        "Start your project 5",
-        "Test",
-        "Wikihouse production checklist",
-        "Your next step",
-      ],
-      logo: "https://raw.githubusercontent.com/theopensystemslab/planx-team-logos/main/wikihouse.svg",
-      favicon: null,
-      primaryColour: "#000000",
-      linkColour: "#000000",
-      actionColour: "#000000",
-    },
-  ];
+  const adminPanelData: AdminPanelData[] | undefined = useStore(
+    (state) => state.adminPanelData,
+  );
 
   const teamsToFilterOut = [
     "WikiHouse",
@@ -1359,10 +34,6 @@ export const PlatformAdminPanel = () => {
     (team) => !teamsToFilterOut.includes(team.name),
   );
 
-  const baseColDef: Partial<GridColDef> = {
-    width: 150,
-  };
-
   const renderIsItemConfigured = (
     params: GridRenderCellParams<any, any, any, GridTreeNodeWithRender>,
   ) => {
@@ -1371,18 +42,15 @@ export const PlatformAdminPanel = () => {
 
   const columns: GridColDef[] = [
     {
-      ...baseColDef,
       field: "name",
       headerName: "Team",
     },
     {
-      ...baseColDef,
       field: "referenceCode",
       headerName: "Reference code",
     },
 
     {
-      ...baseColDef,
       field: "liveFlows",
       headerName: "Live services",
       width: 450,
@@ -1411,13 +79,11 @@ export const PlatformAdminPanel = () => {
       },
     },
     {
-      ...baseColDef,
       field: "planningDataEnabled",
       headerName: "Planning constraints",
       renderCell: renderIsItemConfigured,
     },
     {
-      ...baseColDef,
       field: "article4sEnabled",
       headerName: "Article 4s (API)",
       renderCell: (params) => {
@@ -1425,13 +91,11 @@ export const PlatformAdminPanel = () => {
       },
     },
     {
-      ...baseColDef,
       field: "govpayEnabled",
       headerName: "GOV.UK Pay",
       renderCell: renderIsItemConfigured,
     },
     {
-      ...baseColDef,
       field: "govnotifyPersonalisation",
       headerName: "GOV.UK Notify",
       renderCell: (params) => {
@@ -1439,37 +103,31 @@ export const PlatformAdminPanel = () => {
       },
     },
     {
-      ...baseColDef,
       field: "sendToEmailAddress",
       headerName: "Send to email",
       renderCell: renderIsItemConfigured,
     },
     {
-      ...baseColDef,
       field: "bopsSubmissionURL",
       headerName: "BOPS",
       renderCell: renderIsItemConfigured,
     },
     {
-      ...baseColDef,
       field: "powerAutomateEnabled",
       headerName: "Power automate",
       renderCell: renderIsItemConfigured,
     },
     {
-      ...baseColDef,
       field: "subdomain",
       headerName: "Subdomain",
       renderCell: renderIsItemConfigured,
     },
     {
-      ...baseColDef,
       field: "logo",
       headerName: "Logo",
       renderCell: renderIsItemConfigured,
     },
     {
-      ...baseColDef,
       field: "favicon",
       headerName: "Favicon",
       renderCell: renderIsItemConfigured,
@@ -1489,18 +147,7 @@ export const PlatformAdminPanel = () => {
         </Typography>
       </SettingsSection>
       <SettingsSection>
-        <Box sx={{ height: "100vh", flex: 1, position: "relative" }}>
-          <Box sx={{ inset: 0, position: "absolute" }}>
-            <DataGrid
-              rows={filteredPanelData}
-              columns={columns}
-              getRowHeight={() => "auto"}
-              getRowClassName={(params) =>
-                params.indexRelativeToCurrentPage % 2 === 0 ? "even" : "odd"
-              }
-            />
-          </Box>
-        </Box>
+        <DataTable rows={filteredPanelData} columns={columns} />
       </SettingsSection>
     </Container>
   );
